refactor: Checkbox 외부 API를 props 형태로 전환
