+++ packages/jds/src/components/Checkbox/Checkbox.stories.tsx
@@ -4,18 +4,19 @@ import { useState } from "react";
 
 import { Checkbox } from "./Checkbox";
 import { CHECKBOX_SIZE_OPTIONS, CHECKBOX_VARIANT_OPTIONS, type CheckedState } from "./checkbox.types";
+import { CheckboxGroup } from "./CheckboxGroup";
 
-const meta: Meta<typeof Checkbox.Item> = {
+const meta = {
   title: "Components/Checkbox",
-  component: Checkbox.Item,
+  component: Checkbox,
   parameters: {
     layout: "centered",
   },
-} satisfies Meta<typeof Checkbox.Item>;
+} satisfies Meta<typeof Checkbox>;
 
 export default meta;
 
-type Story = StoryObj<typeof Checkbox.Item>;
+type Story = StoryObj<typeof Checkbox>;
 
 const GROUP_GAP = "12px";
 const SECTION_GAP = "32px";
@@ -23,10 +24,10 @@ const SECTION_GAP = "32px";
 export const CheckboxBasicSizes: Story = {
   render: () => (
     <FlexRow>
-      <Checkbox.Control size='lg' />
-      <Checkbox.Control size='md' />
-      <Checkbox.Control size='sm' />
-      <Checkbox.Control size='xs' />
+      <Checkbox size='lg' />
+      <Checkbox size='md' />
+      <Checkbox size='sm' />
+      <Checkbox size='xs' />
     </FlexRow>
   ),
 };
@@ -35,16 +36,16 @@ export const CheckboxBasicStates: Story = {
   render: () => (
     <FlexColumn>
       <FlexRow>
-        <Checkbox.Control checked={false} onCheckedChange={() => {}} />
-        <Checkbox.Control checked={true} onCheckedChange={() => {}} />
-        <Checkbox.Control checked='indeterminate' onCheckedChange={() => {}} />
-        <Checkbox.Control checked={false} isInvalid onCheckedChange={() => {}} />
+        <Checkbox checked={false} onCheckedChange={() => {}} />
+        <Checkbox checked={true} onCheckedChange={() => {}} />
+        <Checkbox checked='indeterminate' onCheckedChange={() => {}} />
+        <Checkbox checked={false} isInvalid onCheckedChange={() => {}} />
       </FlexRow>
       <FlexRow>
-        <Checkbox.Control checked={false} disabled onCheckedChange={() => {}} />
-        <Checkbox.Control checked={true} disabled onCheckedChange={() => {}} />
-        <Checkbox.Control checked='indeterminate' disabled onCheckedChange={() => {}} />
-        <Checkbox.Control checked={false} isInvalid disabled onCheckedChange={() => {}} />
+        <Checkbox checked={false} disabled onCheckedChange={() => {}} />
+        <Checkbox checked={true} disabled onCheckedChange={() => {}} />
+        <Checkbox checked='indeterminate' disabled onCheckedChange={() => {}} />
+        <Checkbox checked={false} isInvalid disabled onCheckedChange={() => {}} />
       </FlexRow>
     </FlexColumn>
   ),
@@ -55,19 +56,12 @@ export const CheckboxItemStyle: Story = {
     <FlexColumn gap={SECTION_GAP}>
       <FlexColumn gap={GROUP_GAP}>
         {CHECKBOX_SIZE_OPTIONS.map(size => (
-          <Checkbox.Item key={size} size={size} variant='hollow'>
-            <Checkbox.Control value='item' />
-            <Checkbox.Label>레이블</Checkbox.Label>
-          </Checkbox.Item>
+          <Checkbox key={size} size={size} variant='hollow' label='레이블' />
         ))}
       </FlexColumn>
       <FlexColumn gap={GROUP_GAP}>
         {CHECKBOX_SIZE_OPTIONS.map(size => (
-          <Checkbox.Item key={size} size={size} variant='outlined'>
-            <Checkbox.Control value='item' />
-            <Checkbox.Label>레이블</Checkbox.Label>
-            <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-          </Checkbox.Item>
+          <Checkbox key={size} size={size} variant='outlined' label='레이블' helper='헬퍼 텍스트' />
         ))}
       </FlexColumn>
     </FlexColumn>
@@ -77,15 +71,8 @@ export const CheckboxItemStyle: Story = {
 export const CheckboxItemDisabled: Story = {
   render: () => (
     <FlexColumn gap={GROUP_GAP}>
-      <Checkbox.Item variant='hollow' disabled>
-        <Checkbox.Control value='item' />
-        <Checkbox.Label>레이블</Checkbox.Label>
-      </Checkbox.Item>
-      <Checkbox.Item variant='outlined' disabled>
-        <Checkbox.Control value='item' />
-        <Checkbox.Label>레이블</Checkbox.Label>
-        <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-      </Checkbox.Item>
+      <Checkbox variant='hollow' disabled label='레이블' />
+      <Checkbox variant='outlined' disabled label='레이블' helper='헬퍼 텍스트' />
     </FlexColumn>
   ),
 };
@@ -93,25 +80,10 @@ export const CheckboxItemDisabled: Story = {
 export const CheckboxItemInvalid: Story = {
   render: () => (
     <FlexColumn gap={GROUP_GAP}>
-      <Checkbox.Item variant='hollow' isInvalid>
-        <Checkbox.Control value='item' />
-        <Checkbox.Label>레이블</Checkbox.Label>
-      </Checkbox.Item>
-      <Checkbox.Item variant='outlined' isInvalid>
-        <Checkbox.Control value='item' />
-        <Checkbox.Label>레이블</Checkbox.Label>
-        <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-      </Checkbox.Item>
-      <Checkbox.Item variant='outlined' isInvalid disabled>
-        <Checkbox.Control value='item' />
-        <Checkbox.Label>레이블</Checkbox.Label>
-        <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-      </Checkbox.Item>
-      <Checkbox.Item variant='outlined' isInvalid>
-        <Checkbox.Control value='item' defaultChecked />
-        <Checkbox.Label>레이블</Checkbox.Label>
-        <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-      </Checkbox.Item>
+      <Checkbox variant='hollow' isInvalid label='레이블' />
+      <Checkbox variant='outlined' isInvalid label='레이블' helper='헬퍼 텍스트' />
+      <Checkbox variant='outlined' isInvalid disabled label='레이블' helper='헬퍼 텍스트' />
+      <Checkbox variant='outlined' isInvalid defaultChecked label='레이블' helper='헬퍼 텍스트' />
     </FlexColumn>
   ),
   parameters: {
@@ -127,95 +99,69 @@ export const CheckboxItemInvalid: Story = {
 export const CheckboxGroupUncontrolled: Story = {
   render: () => (
     <FlexColumn gap={GROUP_GAP}>
-      <Checkbox.Root defaultValue={["2"]} name='groupUncontrolled'>
-        <Checkbox.Item>
-          <Checkbox.Control value='1' />
-          <Checkbox.Label>레이블</Checkbox.Label>
-        </Checkbox.Item>
-        <Checkbox.Item>
-          <Checkbox.Control value='2' />
-          <Checkbox.Label>레이블 (기본 선택)</Checkbox.Label>
-        </Checkbox.Item>
-        <Checkbox.Item>
-          <Checkbox.Control value='3' />
-          <Checkbox.Label>레이블</Checkbox.Label>
-        </Checkbox.Item>
-      </Checkbox.Root>
+      <CheckboxGroup
+        defaultValue={["2"]}
+        name='groupUncontrolled'
+        options={[
+          { value: "1", label: "레이블" },
+          { value: "2", label: "레이블 (기본 선택)" },
+          { value: "3", label: "레이블" },
+        ]}
+      />
     </FlexColumn>
   ),
 };
 
 export const CheckboxGroupControlled: Story = {
-  render: () => {
-    const ControlledGroup = () => {
-      const [selected, setSelected] = useState<string[]>(["1"]);
+  render: function Render() {
+    const [selected, setSelected] = useState<string[]>(["1"]);
 
-      return (
-        <FlexColumn>
-          <span>선택: {selected.join(", ") || "(없음)"}</span>
-          <FlexColumn gap={GROUP_GAP}>
-            <Checkbox.Root value={selected} onChange={setSelected} name='groupControlled'>
-              <Checkbox.Item>
-                <Checkbox.Control value='1' />
-                <Checkbox.Label>레이블</Checkbox.Label>
-              </Checkbox.Item>
-              <Checkbox.Item>
-                <Checkbox.Control value='2' />
-                <Checkbox.Label>레이블</Checkbox.Label>
-              </Checkbox.Item>
-              <Checkbox.Item>
-                <Checkbox.Control value='3' />
-                <Checkbox.Label>레이블</Checkbox.Label>
-              </Checkbox.Item>
-            </Checkbox.Root>
-          </FlexColumn>
+    return (
+      <FlexColumn>
+        <span>선택: {selected.join(", ") || "(없음)"}</span>
+        <FlexColumn gap={GROUP_GAP}>
+          <CheckboxGroup
+            value={selected}
+            onChange={setSelected}
+            name='groupControlled'
+            options={[
+              { value: "1", label: "레이블" },
+              { value: "2", label: "레이블" },
+              { value: "3", label: "레이블" },
+            ]}
+          />
         </FlexColumn>
-      );
-    };
-
-    return <ControlledGroup />;
+      </FlexColumn>
+    );
   },
 };
 
 export const CheckboxGroupSelectAll: Story = {
-  render: () => {
+  render: function Render() {
     const ALL = ["1", "2", "3"];
+    const [selected, setSelected] = useState<string[]>(["1"]);
 
-    const SelectAllGroup = () => {
-      const [selected, setSelected] = useState<string[]>(["1"]);
+    const isAllChecked = selected.length === ALL.length;
+    const isSomeChecked = selected.length > 0 && !isAllChecked;
+    const parentState: CheckedState = isAllChecked ? true : isSomeChecked ? "indeterminate" : false;
 
-      const isAllChecked = selected.length === ALL.length;
-      const isSomeChecked = selected.length > 0 && !isAllChecked;
-      const parentState: CheckedState = isAllChecked
-        ? true
-        : isSomeChecked
-          ? "indeterminate"
-          : false;
-
-      return (
+    return (
+      <FlexColumn gap={GROUP_GAP}>
+        <Checkbox
+          checked={parentState}
+          onCheckedChange={() => setSelected(isAllChecked ? [] : [...ALL])}
+          label='전체 선택'
+        />
         <FlexColumn gap={GROUP_GAP}>
-          <Checkbox.Item>
-            <Checkbox.Control
-              checked={parentState}
-              onCheckedChange={() => setSelected(isAllChecked ? [] : [...ALL])}
-            />
-            <Checkbox.Label>전체 선택</Checkbox.Label>
-          </Checkbox.Item>
-          <FlexColumn gap={GROUP_GAP}>
-            <Checkbox.Root value={selected} onChange={setSelected} name='groupSelectAll'>
-              {ALL.map(value => (
-                <Checkbox.Item key={value}>
-                  <Checkbox.Control value={value} />
-                  <Checkbox.Label>레이블 {value}</Checkbox.Label>
-                </Checkbox.Item>
-              ))}
-            </Checkbox.Root>
-          </FlexColumn>
+          <CheckboxGroup
+            value={selected}
+            onChange={setSelected}
+            name='groupSelectAll'
+            options={ALL.map(value => ({ value, label: `레이블 ${value}` }))}
+          />
         </FlexColumn>
-      );
-    };
-
-    return <SelectAllGroup />;
+      </FlexColumn>
+    );
   },
   parameters: {
     docs: {
@@ -230,56 +176,45 @@ export const CheckboxGroupSelectAll: Story = {
 export const CheckboxGroupDisabled: Story = {
   render: () => (
     <FlexColumn gap={GROUP_GAP}>
-      <Checkbox.Root disabled defaultValue={["2"]} name='groupDisabled'>
-        <Checkbox.Item>
-          <Checkbox.Control value='1' />
-          <Checkbox.Label>레이블</Checkbox.Label>
-        </Checkbox.Item>
-        <Checkbox.Item>
-          <Checkbox.Control value='2' />
-          <Checkbox.Label>레이블</Checkbox.Label>
-        </Checkbox.Item>
-      </Checkbox.Root>
+      <CheckboxGroup
+        disabled
+        defaultValue={["2"]}
+        name='groupDisabled'
+        options={[
+          { value: "1", label: "레이블" },
+          { value: "2", label: "레이블" },
+        ]}
+      />
     </FlexColumn>
   ),
 };
 
 export const CheckboxGroupInvalid: Story = {
-  render: () => {
-    const InvalidGroup = () => {
-      const [selected, setSelected] = useState<string[]>([]);
-      const isInvalid = selected.length === 0;
+  render: function Render() {
+    const [selected, setSelected] = useState<string[]>([]);
+    const isInvalid = selected.length === 0;
 
-      return (
-        <FlexColumn gap={GROUP_GAP}>
-          <Checkbox.Root
-            value={selected}
-            onChange={setSelected}
-            isInvalid={isInvalid}
-            name='groupInvalid'
-            variant='outlined'
-          >
-            <Checkbox.Item>
-              <Checkbox.Control value='1' />
-              <Checkbox.Label>레이블</Checkbox.Label>
-              <Checkbox.Helper>1개 이상 선택해 주세요</Checkbox.Helper>
-            </Checkbox.Item>
-            <Checkbox.Item>
-              <Checkbox.Control value='2' />
-              <Checkbox.Label>레이블</Checkbox.Label>
-            </Checkbox.Item>
-          </Checkbox.Root>
-        </FlexColumn>
-      );
-    };
-
-    return <InvalidGroup />;
+    return (
+      <FlexColumn gap={GROUP_GAP}>
+        <CheckboxGroup
+          value={selected}
+          onChange={setSelected}
+          isInvalid={isInvalid}
+          name='groupInvalid'
+          variant='outlined'
+          options={[
+            { value: "1", label: "레이블", helper: "1개 이상 선택해 주세요" },
+            { value: "2", label: "레이블" },
+          ]}
+        />
+      </FlexColumn>
+    );
   },
   parameters: {
     docs: {
       description: {
         story:
-          "`Checkbox.Root`의 `isInvalid`를 선택 상태에서 파생하면 그룹 단위 유효성 검사를 표현할 수 있습니다. 아무것도 선택하지 않은 초기 상태에서는 invalid가 그룹 전체로 전파되며, 하나 이상 선택하면 invalid가 해제됩니다.",
+          "`CheckboxGroup`의 `isInvalid`를 선택 상태에서 파생하면 그룹 단위 유효성 검사를 표현할 수 있습니다. 아무것도 선택하지 않은 초기 상태에서는 invalid가 그룹 전체로 전파되며, 하나 이상 선택하면 invalid가 해제됩니다.",
       },
     },
   },
@@ -291,38 +226,21 @@ export const CheckboxComprehensiveMatrix: Story = {
       <FlexColumn>
         {CHECKBOX_SIZE_OPTIONS.map(size => (
           <FlexRow key={size}>
-            <Checkbox.Control size={size} checked={false} onCheckedChange={() => {}} />
-            <Checkbox.Control size={size} checked={true} onCheckedChange={() => {}} />
-            <Checkbox.Control size={size} checked='indeterminate' onCheckedChange={() => {}} />
-            <Checkbox.Control size={size} checked={false} disabled onCheckedChange={() => {}} />
-            <Checkbox.Control size={size} checked={true} disabled onCheckedChange={() => {}} />
-            <Checkbox.Control
-              size={size}
-              checked='indeterminate'
-              disabled
-              onCheckedChange={() => {}}
-            />
+            <Checkbox size={size} checked={false} onCheckedChange={() => {}} />
+            <Checkbox size={size} checked={true} onCheckedChange={() => {}} />
+            <Checkbox size={size} checked='indeterminate' onCheckedChange={() => {}} />
+            <Checkbox size={size} checked={false} disabled onCheckedChange={() => {}} />
+            <Checkbox size={size} checked={true} disabled onCheckedChange={() => {}} />
+            <Checkbox size={size} checked='indeterminate' disabled onCheckedChange={() => {}} />
           </FlexRow>
         ))}
       </FlexColumn>
       <FlexColumn gap={SECTION_GAP}>
         {CHECKBOX_VARIANT_OPTIONS.map(variant => (
           <FlexColumn key={variant} gap={GROUP_GAP}>
-            <Checkbox.Item variant={variant}>
-              <Checkbox.Control defaultChecked />
-              <Checkbox.Label>레이블</Checkbox.Label>
-              <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-            </Checkbox.Item>
-            <Checkbox.Item variant={variant} disabled>
-              <Checkbox.Control />
-              <Checkbox.Label>레이블</Checkbox.Label>
-              <Checkbox.Helper>헬퍼 텍스트</Checkbox.Helper>
-            </Checkbox.Item>
-            <Checkbox.Item variant={variant} isInvalid>
-              <Checkbox.Control />
-              <Checkbox.Label>레이블</Checkbox.Label>
-              <Checkbox.Helper>필수 항목입니다</Checkbox.Helper>
-            </Checkbox.Item>
+            <Checkbox variant={variant} defaultChecked label='레이블' helper='헬퍼 텍스트' />
+            <Checkbox variant={variant} disabled label='레이블' helper='헬퍼 텍스트' />
+            <Checkbox variant={variant} isInvalid label='레이블' helper='필수 항목입니다' />
           </FlexColumn>
         ))}
       </FlexColumn>
